Add changes for part 2

+++ Client/Assets/Scripts/UIManager.cs
@@ -45,7 +45,7 @@ public class UIManager : MonoBehaviour
 
     public void SendName()
     {
-        Message message = Message.Create(MessageSendMode.reliable, (ushort)ClientToServerId.name);
+        Message message = Message.Create(MessageSendMode.reliable, ClientToServerId.name);
         message.AddString(usernameField.text);
         NetworkManager.Singleton.Client.Send(message);
     }
